+++ sources/server/web/rest/service/Maintenance.h
@@ -7,6 +7,7 @@
 #include "database/IDatabaseRequester.h"
 #include "database/IDataProvider.h"
 #include "database/IKeywordRequester.h"
+#include "IUploadFileManager.h"
 
 namespace web
 {
@@ -19,7 +20,8 @@ namespace web
          public:
             explicit CMaintenance(boost::shared_ptr<const IPathProvider> pathProvider,
                                   const boost::shared_ptr<database::IDataProvider>& dataProvider,
-                                  boost::shared_ptr<task::CScheduler> taskScheduler);
+                                  boost::shared_ptr<task::CScheduler> taskScheduler,
+                                  boost::shared_ptr<IUploadFileManager> uploadFileManager);
             virtual ~CMaintenance() = default;
 
             // IRestService implementation
@@ -58,6 +60,11 @@ namespace web
                                                                                             const std::vector<std::string>& parameters,
                                                                                             const std::string& requestContent) const;
 
+            boost::shared_ptr<std::string> fileUploadChunkRead(const std::string& requestContent) const;
+            std::string fileUploadChunkReadGuid(const std::string& requestContent) const;
+            std::string fileUploadChunkReadFilename(const std::string& requestContent) const;
+            unsigned int fileUploadChunkReadFileSize(const std::string& requestContent) const;
+
             static std::string m_restKeyword;
             boost::shared_ptr<const IPathProvider> m_pathProvider;
             boost::shared_ptr<database::IDataProvider> m_dataProvider;
@@ -65,6 +72,7 @@ namespace web
             boost::shared_ptr<database::IKeywordRequester> m_keywordRequester;
             boost::shared_ptr<database::IAcquisitionRequester> m_acquisitionRequester;
             boost::shared_ptr<task::CScheduler> m_taskScheduler;
+            boost::shared_ptr<IUploadFileManager> m_uploadFileManager;
          };
       } //namespace service
    } //namespace rest